fix config loading when repo is under a symlinked directory
Say we have a directory in /data/foo/src, which is symlinked from
/home/foo/src. There are .mrconfig files in /home/foo and /home/foo/src,
and we work in a repo in /home/foo/src/repo.

Currently, if we run mr from /home/foo/src/repo, the config loading goes
as follows:

1. load config at /home/foo/.rmconfig
2. load config at /data/foo/src/.mrconfig
3. if no DEFAULT section is defined in /data/foo/src/.mrconfig, climb up
   path up to the root and grab the first DEFAULT section found in a
   .mrconfig file.
4. As a last resort, use DEFAULT embedded in the perl file.

Since the real path of our repo is not contained in the home directory,
/home/foo/.mrconfig loaded, but is never reached to grab its DEFAULT
section.

This breaks the expectation that /home/foo/.mrconfig applies globally
for a given user, irrespective which is its working directory. We fix it
by making sure we always look for a DEFAULT section in
/home/foo/.mrconfig between steps 3 and 4 above.